Renamed package org.odlabs to org.wicketstuff
In order to make clear that this is a fork of the original, and because
we have the opportunity to rename a package (this is a new major
version after all), we renamed the package from org.odlabs to
org.wicketstuff.

While we don't think the original will receive any maintenance in the
future, this makes sure that the original developers are no longer 'on
the hook' for releases we publish from the Wicket Stuff organization.